+++ src/main/java/gr/codehub/accenture/technicoproject/repository/PropertyOwnerRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PropertyOwnerRepository extends JpaRepository<PropertyOwner, Long> {
+public interface PropertyOwnerRepository extends JpaRepository<PropertyOwner, Integer> {
 }
