+++ src/main/java/gr/codehub/accenture/technicoproject/repository/PropertyRepairOrderRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PropertyRepairOrderRepository extends JpaRepository<PropertyRepairOrder, Long> {
+public interface PropertyRepairOrderRepository extends JpaRepository<PropertyRepairOrder, Integer> {
 }
